Update RBush due to security
RBush 3.2.0 depends on netstandard 1.2, which transitively depends
classic vulnerabel packages RegEx, HttpClient, ect.

I am pretty sure it's a false positive due to poor implementation of
analyzer (Visual Studio reports it for netstandard2.1, but not for
netstandard 2.0), but it's just not worth to try to fight the tide.

I don't think these vulnerable packages are not actually used on any
maintained runtime. The runtime will just load the implementation
shipped with maintained runtime that have this fixed (at least that
was a response of netstandard team and I checked it for NetFx and
Core).

Just another consequence of not separating implementation and contract
as nuget packages for netstandard...